Implement modified sieve
The original sieving method does not provide prevention against prime
number that is smaller than the cluster size. The reason of it in the
original paper is that shamir secret sharing is not possible with M
having prime factors smaller than cluster size.

This modified version ignores this limitation and directly use all prime
numbers smaller than the max sieve bound to compute M. It has no
observerable side effects so far and massively improved performance.

We are still not clear about the limitation written in the original
paper, just as we do not understand how in the paper they "fix" the
small prime number problems. We will continue to do research on it but
this improvement made everything usable.